# api/docs/swagger.yaml
components:
    schemas:
...
        ErrorResponse:
            properties:
                code:
                    example: InternalError
                    type: string
                error:
                    example: An unexpected error occurred
                    type: string
            type: object
...
        NoteIn:
            properties:
                content:
                    example: Notes from the meeting with the client.
                    type: string
                title:
                    example: Meeting Notes
                    type: string
            required:
                - content
                - title
            type: object
        NoteOut:
            properties:
                archived:
                    type: boolean
                attachments:
                    items:
                        $ref: '#/components/schemas/AttachmentOut'
                    type: array
                content:
                    example: Notes from the meeting with the client.
                    type: string
                created_at:
                    type: string
                encrypted:
                    type: boolean
                id:
                    example: 123e4567-e89b-12d3-a456-426614174000
                    type: string
                title:
                    example: Meeting Notes
                    type: string
                updated_at:
                    type: string
            type: object
        PresignDownloadResponse:
            properties:
                url:
                    type: string
            type: object
        PresignUploadRequest:
            properties:
                content_type:
                    example: text/plain
                    type: string
                filename:
                    example: example.txt
                    type: string
            required:
                - content_type
                - filename
            type: object
        PresignUploadResponse:
            properties:
                key:
                    example: attachments/123e4567-e89b-12d3-a456-426614174000/example.txt
                    type: string
                url:
                    example: https://s3.com/upload?key=example.txt
                    type: string
            type: object
...
        ShareToUserRequest:
            properties:
                permission:
                    description: '"read" or "write"'
                    example: read
                    type: string
                shared_with:
                    example: 123e4567-e89b-12d3-a456-426614174000
                    type: string
            required:
                - permission
                - shared_with
            type: object
...
info:
    contact:
        name: Kit
        url: https://github.com/kit-g
    description: A simple Secrets storage API
    license:
        name: MIT
        url: https://opensource.org/licenses/MIT
    termsOfService: http://swagger.io/terms/
    title: Vault API
    version: "1.0"
openapi: 3.0.3
paths:
...
    /notes:
        get:
            description: Returns paginated notes for the authenticated user with optional filtering
            operationId: getNotes
            parameters:
                - description: Page number
                  in: query
                  name: page
                  schema:
                    default: 1
                    type: integer
                - description: Items per page
                  in: query
                  name: limit
                  schema:
                    default: 10
                    type: integer
                - description: Filter by archived status
                  in: query
                  name: archived
                  schema:
                    type: boolean
                - description: Filter by encrypted status
                  in: query
                  name: encrypted
                  schema:
                    type: boolean
            responses:
                "200":
                    content:
                        application/json:
                            schema:
                                items:
                                    $ref: '#/components/schemas/NoteOut'
                                type: array
                    description: OK
                "401":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ErrorResponse'
                    description: Unauthorized
                "500":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ErrorResponse'
                    description: Server error
            security:
                - BearerAuth: []
            summary: List user notes
            tags:
                - notes
        post:
            description: Creates a note for the authenticated user
            operationId: createNote
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/NoteIn'
                description: Note object
                required: true
                x-originalParamName: note
            responses:
                "201":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/NoteOut'
                    description: Created
                "400":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ErrorResponse'
                    description: Bad Request
                "401":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ErrorResponse'
                    description: Unauthorized
                "500":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ErrorResponse'
                    description: Internal Server Error
            security:
                - BearerAuth: []
            summary: Create a new note
            tags:
                - notes
    /notes/{noteId}:
        delete:
            description: Deletes a note owned by the authenticated user
            operationId: deleteNote
            parameters:
                - description: Note ID
                  in: path
                  name: noteId
                  required: true
                  schema:
                    type: string
                - description: Hard delete flag
                  in: query
                  name: hard
                  schema:
                    default: false
                    type: boolean
            responses:
                "204":
                    description: No Content
                "400":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ErrorResponse'
                    description: Bad Request
                "401":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ErrorResponse'
                    description: Unauthorized
                "404":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ErrorResponse'
                    description: Not Found
                "500":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ErrorResponse'
                    description: Internal Server Error
            security:
                - BearerAuth: []
            summary: Delete a note
            tags:
                - notes
        get:
            description: Retrieves a specific note owned by the authenticated user
            operationId: getNote
            parameters:
                - description: Note UUID
                  in: path
                  name: noteId
                  required: true
                  schema:
                    type: string
            responses:
                "200":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/NoteOut'
                    description: OK
                "400":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ErrorResponse'
                    description: Bad Request
                "401":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ErrorResponse'
                    description: Unauthorized
                "404":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ErrorResponse'
                    description: Not Found
                "500":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ErrorResponse'
                    description: Internal Server Error
            security:
                - BearerAuth: []
            summary: Get a single note
            tags:
                - notes
        put:
            description: Updates the note fields for the authenticated user
            operationId: editNote
            parameters:
                - description: Note ID
                  in: path
                  name: noteId
                  required: true
                  schema:
                    type: string
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/NoteIn'
                description: Note fields
                required: true
                x-originalParamName: note
            responses:
                "200":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/NoteOut'
                    description: OK
                "400":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ErrorResponse'
                    description: Bad Request
                "401":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ErrorResponse'
                    description: Unauthorized
                "404":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ErrorResponse'
                    description: Not Found
                "500":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ErrorResponse'
                    description: Internal Server Error
            security:
                - BearerAuth: []
            summary: Edit a note
            tags:
                - notes
    /notes/{noteId}/attachments:
        post:
            description: Generates a presigned URL for uploading an attachment to a specific note
            operationId: getUploadURL
            parameters:
                - description: Note ID
                  in: path
                  name: noteId
                  required: true
                  schema:
                    type: string
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/PresignUploadRequest'
                description: Upload parameters
                required: true
                x-originalParamName: body
            responses:
                "200":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/PresignUploadResponse'
                    description: OK
                "400":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ErrorResponse'
                    description: Bad Request
                "401":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ErrorResponse'
                    description: Unauthorized
                "500":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ErrorResponse'
                    description: Internal Server Error
            summary: Generate a presigned S3 upload URL
            tags:
                - notes
    /notes/{noteId}/attachments/{attachmentId}:
        delete:
            operationId: deleteAttachment
            parameters:
                - description: Note ID
                  in: path
                  name: noteId
                  required: true
                  schema:
                    type: string
                - description: Attachment ID
                  in: path
                  name: attachmentId
                  required: true
                  schema:
                    type: string
            responses:
                "204":
                    description: No Content
                "400":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ErrorResponse'
                    description: Bad Request
                "401":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ErrorResponse'
                    description: Unauthorized
                "404":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ErrorResponse'
                    description: Not Found
            security:
                - BearerAuth: []
            summary: Delete an attachment
            tags:
                - notes
        get:
            description: Generates a temporary URL for securely downloading a note's attachment.
            operationId: getDownloadURL
            parameters:
                - description: Note ID (UUID)
                  in: path
                  name: noteId
                  required: true
                  schema:
                    type: string
                - description: Attachment ID (UUID)
                  in: path
                  name: attachmentId
                  required: true
                  schema:
                    type: string
            responses:
                "200":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/PresignDownloadResponse'
                    description: OK
                "400":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ErrorResponse'
                    description: Bad Request
                "401":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ErrorResponse'
                    description: Unauthorized
                "404":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ErrorResponse'
                    description: Not Found
                "500":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ErrorResponse'
                    description: Internal Server Error
            security:
                - BearerAuth: []
            summary: Get presigned download URL for an attachment
            tags:
                - notes
    /notes/{noteId}/restore:
        post:
            description: Restores a soft-deleted note owned by the authenticated user
            operationId: restoreNote
            parameters:
                - description: Note ID
                  in: path
                  name: noteId
                  required: true
                  schema:
                    type: string
            responses:
                "204":
                    description: No Content
                "400":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ErrorResponse'
                    description: Bad Request
                "401":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ErrorResponse'
                    description: Unauthorized
                "404":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ErrorResponse'
                    description: Not Found
                "500":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ErrorResponse'
                    description: Internal Server Error
            security:
                - BearerAuth: []
            summary: Restore a deleted note
            tags:
                - notes
    /notes/{noteId}/share:
        post:
            description: Allows the authenticated user to share a note they own with another user, specifying read or write permissions.
            operationId: shareNoteToUser
            parameters:
                - description: Note ID (UUID)
                  in: path
                  name: noteId
                  required: true
                  schema:
                    type: string
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/ShareToUserRequest'
                description: Sharing request
                required: true
                x-originalParamName: request
            responses:
                "204":
                    description: No Content
                "400":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ErrorResponse'
                    description: Bad request (invalid UUID, payload, or permission)
                "404":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ErrorResponse'
                    description: Note not found or not owned by user
                "500":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ErrorResponse'
                    description: Internal server error
            security:
                - BearerAuth: []
            summary: Share note with user
            tags:
                - notes
...
servers:
    - url: https://localhost:8080/